+++ lib/internal/readline/utils.js
@@ -31,8 +31,8 @@ function CSI(strings, ...args) {
 }
 
 CSI.kEscape = kEscape;
-CSI.kClearToBeginning = CSI`1K`;
-CSI.kClearToEnd = CSI`0K`;
+CSI.kClearToLineBeginning = CSI`1K`;
+CSI.kClearToLineEnd = CSI`0K`;
 CSI.kClearLine = CSI`2K`;
 CSI.kClearScreenDown = CSI`0J`;
 
@@ -445,7 +445,27 @@ function* emitKeys(stream) {
   }
 }
 
+// This runs in O(n log n).
+function commonPrefix(strings) {
+  if (!strings || strings.length === 0) {
+    return '';
+  }
+  if (strings.length === 1) {
+    return strings[0];
+  }
+  const sorted = strings.slice().sort();
+  const min = sorted[0];
+  const max = sorted[sorted.length - 1];
+  for (let i = 0; i < min.length; i++) {
+    if (min[i] !== max[i]) {
+      return min.slice(0, i);
+    }
+  }
+  return min;
+}
+
 module.exports = {
+  commonPrefix,
   emitKeys,
   getStringWidth,
   isFullWidthCodePoint,
