+++ lib/readline.js
@@ -48,6 +48,7 @@ const { validateString } = require('internal/validators');
 const { inspect } = require('internal/util/inspect');
 const EventEmitter = require('events');
 const {
+  commonPrefix,
   CSI,
   emitKeys,
   getStringWidth,
@@ -59,8 +60,8 @@ const {
 const { clearTimeout, setTimeout } = require('timers');
 const {
   kEscape,
-  kClearToBeginning,
-  kClearToEnd,
+  kClearToLineBeginning,
+  kClearToLineEnd,
   kClearLine,
   kClearScreenDown
 } = CSI;
@@ -566,23 +567,6 @@ function handleGroup(self, group, width, maxColumns) {
   }
   self._writeToOutput('\r\n');
 }
-
-function commonPrefix(strings) {
-  if (!strings || strings.length === 0) {
-    return '';
-  }
-  if (strings.length === 1) return strings[0];
-  const sorted = strings.slice().sort();
-  const min = sorted[0];
-  const max = sorted[sorted.length - 1];
-  for (let i = 0, len = min.length; i < len; i++) {
-    if (min[i] !== max[i]) {
-      return min.slice(0, i);
-    }
-  }
-  return min;
-}
-
 
 Interface.prototype._wordLeft = function() {
   if (this.cursor > 0) {
@@ -1268,7 +1252,11 @@ function clearLine(stream, dir, callback) {
     return true;
   }
 
-  const type = dir < 0 ? kClearToBeginning : dir > 0 ? kClearToEnd : kClearLine;
+  const type = dir < 0 ?
+    kClearToLineBeginning :
+    dir > 0 ?
+      kClearToLineEnd :
+      kClearLine;
   return stream.write(type, callback);
 }
