+++ lib/repl.js
@@ -77,6 +77,9 @@ const vm = require('vm');
 const path = require('path');
 const fs = require('fs');
 const { Interface } = require('readline');
+const {
+  commonPrefix
+} = require('internal/readline/utils');
 const { Console } = require('console');
 const CJSModule = require('internal/modules/cjs/loader').Module;
 const domain = require('domain');
@@ -1387,25 +1390,6 @@ function complete(line, callback) {
   }
 }
 
-function longestCommonPrefix(arr = []) {
-  const cnt = arr.length;
-  if (cnt === 0) return '';
-  if (cnt === 1) return arr[0];
-
-  const first = arr[0];
-  // complexity: O(m * n)
-  for (let m = 0; m < first.length; m++) {
-    const c = first[m];
-    for (let n = 1; n < cnt; n++) {
-      const entry = arr[n];
-      if (m >= entry.length || c !== entry[m]) {
-        return first.substring(0, m);
-      }
-    }
-  }
-  return first;
-}
-
 REPLServer.prototype.completeOnEditorMode = (callback) => (err, results) => {
   if (err) return callback(err);
 
@@ -1418,7 +1402,7 @@ REPLServer.prototype.completeOnEditorMode = (callback) => (err, results) => {
   const trimCompleteOnPrefix = (v) => v.substring(prefixLength);
   const data = completions.filter(isNotEmpty).map(trimCompleteOnPrefix);
 
-  callback(null, [[`${completeOn}${longestCommonPrefix(data)}`], completeOn]);
+  callback(null, [[`${completeOn}${commonPrefix(data)}`], completeOn]);
 };
 
 REPLServer.prototype.defineCommand = function(keyword, cmd) {
